Test CRUD operations through APIs.
a. Priliminary setup of docker, virtual env and django application was
completed.
b. The settings.py was updated with application and middleware
information.
c. MySQL database was setup and the models for the application were
migrated.
d. API end points for CRUD operations on the app were defined and
tested.